move fake environment to avoid circular import
